+++ mutatest/__init__.py
@@ -1,6 +1,6 @@
 """Mutation initialization.
 """
-__version__ = "0.5.0"
+__version__ = "0.5.1-dev"
 
 __title__ = "mutatest"
 __description__ = "Python mutation testing."